fix g16 test
